Soft-fix keyboard navigation in contact sheet
This commit fix partially keyboard navigation in contact sheet,
especially tags field that stuck the tabbing flow.
Due to the way the contact sheet forms are built, fix the whole tabbing
flow is a way harder and need a complex revamping of Ui. I suggest to do
it later when reloading the contact interface (which will take a
significant amount of time).