Add front tests stack powered by Mocha
Tests are ran by Mocha, which can be executed headlessly to dislay reports in
console. Unfortunately, it means that app must have a correct dependencies AST,
which is currently impossible when main libs parts (such as Backbone) are
loaded into the globals scope. We'll can handle this when Brunch will load
properly deps from NPM in its CommonJS AST.

In the meantime, we should rely on browser tests environment, so test are
performed in the browser and are available for reports at the `/test/` URI.