Fix tests
Note: Client tests are temporary unavailable due to the lack of support of npm
dependencies in Brunch 1.x. Running client tests in a headless environment and
with npm dependencies with mocha is on its way, but need more work to have it
available and depends of other tasks (such as the node v4 support).