Allows to disable sorting, improve card processing
Contact sorting is not always useful and is very resource consuming. This
commit allows to disable sorting (returned result of comparison is always -1).
It comes with a proper way to deal with vcard import. It creates contacts
one by one. That way contact creation is more properly handled. Bulk creation
could make it even faster.