Add mediatype to datapoints
This step is the first in enabling the sharing of files between Cozy
instances: we need a reliable way - on the server side - to extract the
url of a Cozy cloud from a contact.

The solution we chose was to add a mediatype to the datapoint structure:
we can interpret an url according to the value of its mediatype. For
instance if the mediatype is 'cloud:cozy' we know that the url
represents the address of a Cozy instance and if the mediatype is empty
it's simply an email address.

This solution has the advantages of still being compliant to the vCard
RFC and, normally, being transparent.